Fix flaw in send queue timing logic
The sendqueue, if flushed, would retain timing figures from the just-emptied queue. This would cause new messages to be queued unnecessarily, and also exposed a flaw where for long queue windows, filling the queue after a short period of time may lead to the connection being dropped for excess flood.

Thanks to Sevalecan for helping find the bug.